Removes using, `SlackClientExtensions` -> `IncomingWebHookExtensions`.
